database/v1/etl/exposure/usda/pyproject.toml
[tool.poetry]
name = "scenariomip-pipeline"
description = "Used for running batch process of cmip6 data for climate variables and outputting geotiffs and loading data into database"
authors = ["Eric Colllins <ecol07@uw.edu>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.12.3"
xarray = "^2024.7"
numpy = "^2"
s3fs = "^2024"
rioxarray = "^0.17"
cftime = "^1.6"
psycopg2-binary = "^2.9.9"
boto3 = "^1.35.11"
geopandas = "^1.0.1"
xvec = "^0.3.0"
exactextract = "^0.2.1"
fsspec = "^2024.12.0"
zarr = "^3.0.0"
dask = "^2024.12.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
